requirements.txt
playwright>=1.56.0
playwright-stealth>=2.0.0
pandas>=2.2.3
openpyxl>=3.1.5
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
lxml>=5.1.1
2captcha-python>=1.1.3
requests>=2.31.0